Separate transport release and close actions
This commit is very experimental in its nature. It separates transport
release and close actions. Now it will be possible to release transport
without closing BT socket. The file descriptor will be (?) closed later
when the other end of the connection responds to our release request.
This mechanism should allow proper asynchronous polling (polling in one
thread and releasing transport in the other) on a released but not yet
closed BT socket.