Dynamically manage ALSA namehints for BlueALSA PCMs
Connected BlueALSA PCMs are included in lists of PCMs got from
the ALSA namehints API, for example with "aplay -L".

Requires the user to include a new BlueALSA config file in the
ALSA_CONFIG_PATH environment variable so that libasound treats it
as a gobal configuration.

Unfortunately, the majority of existing applications assume that
the PCM list is static and only read it once on start-up. Some
(e.g. kodi) assume that PCMs are all from soundcards and so
refresh the list only in response to a udev event when a card, eg
USB soundcard is plugged/unplugged. So this change is not as
useful as it should be, but still hopefully can benefit some
users.