Make ALSA capture more like a hw device
A capture PCM stream on a `hw` device always generates audio frames
at its configured rate; there are never any intervals where no
frames are produced while it is still running. A bluetooth audio
stream cannot give such a guarantee, so to have the same behaviour
with the BlueALSA ALSA plugin we have to pad with silence when a
break in the incoming stream occurs. This is particularly important
for applications such as `alsaloop` and `zita-a2j` which must
synchronize the captured BlueALSA stream with a hardware device.