alsa pcm plugin: improve capture start up and delay reporting
For capture PCMs, Bluetooth does not report its delay, so the value
given by the D-Bus interface is not useful to an ALSA PCM. Therefore
report only the delay represented by the samples in the ring buffer;
this is the minimum required to comply with the ALSA specification
and is sufficient to support applications such as `alsaloop` which
require a meaningful delay value to function correctly.

Also discard any stale samples in the FIFO when preparing a capture
PCM to to ensure a fresh start, especially after overrun.