ALSA PCM plugin: default to MONOTONIC timestamps
This change does not affect the function of the plugin. Its only
purpose is to simplify the use of the bluealsa PCM plugin in
combination with the dmix plugin as slaves of a multi plugin; this
is now a usable combination with bluealsad multi-client mode
enabled.

The multi plugin requires all its slaves to use the same timestamp
type. Since dmix uses monotonic type and bluealsa previously used
gettimeofday type they could not be used together in a multi
without overriding the dmix default in the ALSA configuration.