Permit only one active SCO transport on a USB adapter
In practice BlueALSA has never been able to support multiple
simultaneous SCO streams on USB adapters, and this change makes
that fact explicit by refusing to acquire a second stream when
one is already running.

A USB isochronous endpoint does not have sufficient bandwidth to
carry more than one wideband (WBS or SWB) SCO stream at a time,
so for those codecs this change is merely enforcing a system
limit anyway. For CVSD a USB adapter can carry up to three
simultaneous SCO streams. 'btusb' supports this by changing the
endpoint configuration to modify the isochronous bandwidth as it
processes each URB. So there is no way for an application to know
the URB payload size atomically before writing or reading from
the SCO socket. On balance it seems reasonable to limit both CVSD
and wideband to a single active duplex stream in order to ensure
that the bandwidth and timing limitations of USB SCO are met.