Add Reconfigurable property to org.bluealsa.PCM1 interface
Bluez 5.80 broke compatibility with some A2DP sink devices such
that it is no longer possible to modify the codec configuration
after the initial profile connection negotiation. When codec
reconfiguration is attempted the device instead closes the A2DP
profile connection.

For such devices the BlueALSA ALSA PCM plugin currently fails if
the rate or channel count of the application stream differs from
the PCM's configured value but is valid within the codec
capabilities. There is no way to discover if a specific device
will fail, other than trying it to see what happens.

This commit introduces a new boolean property, "Reconfigurable",
to the org.bluealsa.PCM1 interface. This property is made
persistent, and when its value is false BlueALSA will not attempt
to modify the codec configuration.

When a user finds that their device drops the A2DP profile
connection when BlueALSA requests a codec configration change,
they can set this property to false so that no further codec
configuration changes will be attempted for that device.